Break last commit out into its own component
