+++ src/sentry/static/sentry/app/views/releases/releaseOverview.jsx
@@ -3,7 +3,7 @@ import React from 'react';
 import LoadingIndicator from '../../components/loadingIndicator';
 import LoadingError from '../../components/loadingError';
 import IconOpen from '../../icons/icon-open';
-import Avatar from '../../components/avatar';
+import LastCommit from '../../components/lastCommit';
 import IssueList from '../../components/issueList';
 import CommitAuthorStats from '../../components/commitAuthorStats';
 import ReleaseProjectStatSparkline from '../../components/releaseProjectStatSparkline';
@@ -114,29 +114,10 @@ const ReleaseOverview = React.createClass({
     return <div className="box empty">{t('None')}</div>;
   },
 
-  renderMessage(message) {
-    if (!message) {
-      return t('No message provided');
-    }
-
-    if (message.length > 100) {
-      let truncated = message.substr(0, 90);
-      let words = truncated.split(' ');
-      // try to not have elipsis mid-word
-      if (words.length > 1) {
-        words.pop();
-        truncated = words.join(' ');
-      }
-      return truncated + '...';
-    }
-    return message;
-  },
-
   render() {
     let {orgId, projectId, version} = this.props.params;
     let {release} = this.context;
     let lastCommit = release.lastCommit;
-    let commitAuthor = lastCommit && lastCommit.author;
 
     if (this.state.loading) return <LoadingIndicator />;
 
@@ -219,23 +200,7 @@ const ReleaseOverview = React.createClass({
             {hasRepos
               ? <div>
                   {lastCommit &&
-                    <div>
-                      <h6 className="nav-header">Last commit</h6>
-                      <div className="commit">
-                        <div className="commit-avatar">
-                          <Avatar user={commitAuthor || {username: '?'}} />
-                        </div>
-                        <div className="commit-message truncate">
-                          {this.renderMessage(lastCommit.message)}
-                        </div>
-                        <div className="commit-meta">
-                          <strong>
-                            {(commitAuthor && commitAuthor.name) || t('Unknown Author')}
-                          </strong>&nbsp;
-                          <TimeSince date={lastCommit.dateCreated} />
-                        </div>
-                      </div>
-                    </div>}
+                    <LastCommit lastCommit={lastCommit} headerClass="nav-header" />}
                   <CommitAuthorStats
                     orgId={orgId}
                     projectId={projectId}
